Split pre-compiled definitions into .h and .c file
This is so that parallel builds work properly. Otherwise, we'd have to
somehow add a dependency in the makefile from Calculator-definitions.cc
to the various generated headers, which I don't quite understand how to
do.

The naming & location of the generated files was not very carefully
chosen--I'd appreciate it if you'd take a look and let me know where
you'd like them to go.